1234 confirm delete

+++ lib/delete_folder_popup.dart
@@ -17,13 +17,23 @@ class DeleteFolderPopup extends StatefulWidget {
 }
 
 class _DeleteFolderPopupState extends State<DeleteFolderPopup> {
+  bool _showConfirmationInput = false;
+  late TextEditingController _confirmationController;
+  bool _wrongConfirmation = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _confirmationController = TextEditingController();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Center(
       child: Padding(
         padding: const EdgeInsets.all(20),
         child: Hero(
-          tag: "Delete-folder-hero",
+          tag: "Delete-image-hero",
           createRectTween: (begin, end) {
             return CustomRectTween(begin: begin, end: end);
           },
@@ -35,7 +45,7 @@ class _DeleteFolderPopupState extends State<DeleteFolderPopup> {
             ),
             child: SizedBox(
               width: 300,
-              height: 180,
+              height: 200,
               child: SingleChildScrollView(
                 child: Padding(
                   padding: const EdgeInsets.symmetric(horizontal: 25, vertical: 20),
@@ -44,7 +54,7 @@ class _DeleteFolderPopupState extends State<DeleteFolderPopup> {
                       const Align(
                         alignment: Alignment.center,
                         child: Text(
-                          "Delete Folder",
+                          "Delete",
                           style: TextStyle(
                             color: Color(0xff284b63),
                             fontSize: 20,
@@ -52,18 +62,51 @@ class _DeleteFolderPopupState extends State<DeleteFolderPopup> {
                           ),
                         ),
                       ),
-                      const SizedBox(height: 10),
-                      const Align(
-                        alignment: Alignment.center,
-                        child: Text(
-                          "Are you sure you want to delete this file? It cannot be recovered.",
-                          style: TextStyle(
-                            fontSize: 14,
-                            color: Colors.black,
-                          ),
-                          textAlign: TextAlign.center,
+                      if (!_showConfirmationInput)
+                        const Column(
+                          children: [
+                            SizedBox(height: 10),
+                            Align(
+                              alignment: Alignment.center,
+                              child: Text(
+                                "Are you sure you want to delete this file? It cannot be recovered.",
+                                style: TextStyle(
+                                  fontSize: 14,
+                                  color: Colors.black,
+                                ),
+                                textAlign: TextAlign.center,
+                              ),
+                            ),
+                          ],
                         ),
-                      ),
+                      if (_showConfirmationInput)
+                        if (_showConfirmationInput)
+                          TextField(
+                            controller: _confirmationController,
+                            keyboardType: TextInputType.number,
+                            obscureText: true,
+                            cursorColor: Colors.black,
+                            style: const TextStyle(color: Colors.black),
+                            decoration: const InputDecoration(
+                              labelText: "Type '1234' to confirm",
+                              labelStyle: TextStyle(color: Colors.black), // Label color
+                              focusedBorder: UnderlineInputBorder(
+                                borderSide: BorderSide(color: Colors.black), // Underline color
+                              ),
+                            ), // Text color
+                          ),
+                      if (_wrongConfirmation)
+                        const Align(
+                          alignment: Alignment.centerLeft,
+                          child: Text(
+                            "Incorrect! Please try again.",
+                            style: TextStyle(
+                              fontSize: 12,
+                              color: Colors.red,
+                            ),
+                            textAlign: TextAlign.left,
+                          ),
+                        ),
                       const SizedBox(height: 20),
                       Align(
                         alignment: Alignment.center,
@@ -80,9 +123,9 @@ class _DeleteFolderPopupState extends State<DeleteFolderPopup> {
                                   borderRadius: BorderRadius.circular(20),
                                 ),
                               ),
-                              child: const Text(
-                                'CANCEL',
-                                style: TextStyle(
+                              child: Text(
+                                _showConfirmationInput ? 'BACK' : 'CANCEL',
+                                style: const TextStyle(
                                   color: Colors.black,
                                   fontWeight: FontWeight.w600,
                                 ),
@@ -91,25 +134,35 @@ class _DeleteFolderPopupState extends State<DeleteFolderPopup> {
                             const SizedBox(width: 20),
                             ElevatedButton(
                               onPressed: () async {
+                                if (_showConfirmationInput) {
+                                  if (_confirmationController.text == '1234') {
+                                    var crackDB = CrackDB();
 
-                                // Perform deletion logic here
-                                // Example: Delete folder from database
-                                var crackDB = CrackDB();
-                                for(int i = 0; i < widget.trackingNo.length; i++){
-                                  List<CrackInfo> trackInfo = await crackDB.fetchALlCrackTracking(trackingNo: int.parse(widget.trackingNo[i].replaceAll(RegExp(r'[^0-9]'),'')));
-                                  for(int j = 0; j < trackInfo.length; j++){
-                                    List<ImageDB> imageList = await crackDB.getImageOnTrackingNo(trackingNo: trackInfo[j].trackingNo);
-                                    await crackDB.deleteImage(id: trackInfo[j].imageID);
-                                    await crackDB.deletePrediction(id: trackInfo[j].imageID);
-                                    await crackDB.deleteCrackInfo(id: trackInfo[j].imageID);
-                                    await crackDB.deleteBuilding(id: trackInfo[j].buildingID);
-                                    await crackDB.deleteFloor(id: trackInfo[j].floorID);
-                                    await crackDB.deleteRoom(id: trackInfo[j].roomID);
-                                    await File(imageList[j].imagePath).delete();
+                                    for(int i = 0; i < widget.trackingNo.length; i++){
+                                      List<CrackInfo> trackInfo = await crackDB.fetchALlCrackTracking(trackingNo: int.parse(widget.trackingNo[i].replaceAll(RegExp(r'[^0-9]'),'')));
+                                      for(int j = 0; j < trackInfo.length; j++){
+                                        List<ImageDB> imageList = await crackDB.getImageOnTrackingNo(trackingNo: trackInfo[j].trackingNo);
+                                        await crackDB.deleteImage(id: trackInfo[j].imageID);
+                                        await crackDB.deletePrediction(id: trackInfo[j].imageID);
+                                        await crackDB.deleteCrackInfo(id: trackInfo[j].imageID);
+                                        await crackDB.deleteBuilding(id: trackInfo[j].buildingID);
+                                        await crackDB.deleteFloor(id: trackInfo[j].floorID);
+                                        await crackDB.deleteRoom(id: trackInfo[j].roomID);
+                                        await File(imageList[j].imagePath).delete();
+                                      }
+                                    }
+
+                                    Navigator.pop(context);
+                                  } else {
+                                    setState(() {
+                                      _wrongConfirmation = true;
+                                    });
                                   }
+                                } else {
+                                  setState(() {
+                                    _showConfirmationInput = true;
+                                  });
                                 }
-
-                                Navigator.pop(context);
                               },
                               style: ElevatedButton.styleFrom(
                                 backgroundColor: const Color(0xff284b63),
@@ -117,9 +170,9 @@ class _DeleteFolderPopupState extends State<DeleteFolderPopup> {
                                   borderRadius: BorderRadius.circular(20),
                                 ),
                               ),
-                              child: const Text(
-                                'CONFIRM',
-                                style: TextStyle(
+                              child: Text(
+                                _showConfirmationInput ? 'CONFIRM' : 'CONFIRM',
+                                style: const TextStyle(
                                   color: Colors.white,
                                   fontWeight: FontWeight.bold,
                                 ),
